refactor: fold the never-shipped merge intermediate into v1
Same reasoning as the path_query_push fold: the direction-agreement
merge landed gated to GROVE_V4 as slot 1, V4 has not been released, and
the per-instance lift took the V4 slot as 2 — leaving slot 1 a dead
dispatch arm no version table selects. Since the slot values are
internal until V4 ships, collapse: the direction rules + limit lift ARE
merge v1 now, GROVE_V4 maps path_query_methods.merge back to 1, and the
dispatcher (and its fail-closed validator) know [0, 1]. Table docs,
tests, and the book chapter renumbered to match.
