+++ grovedb/src/query/merge/v1.rs
@@ -1,23 +1,36 @@
-//! `PathQuery::merge` — **v1** (direction-aware; never shipped alone,
-//! superseded by [`super::v2`] before `GROVE_V4` released).
+//! `PathQuery::merge` — **v1** (`GROVE_V4`+).
 //!
-//! Every input must agree on `left_to_right` — a conflict is a typed
-//! error instead of v0's silent drop — and the shared direction
-//! propagates to the merged root through
-//! [`Query::merge_multiple_directional`] and the final assignment
-//! below, so sub-level inputs no longer lose it under a synthesized
-//! root. Limits and offsets are still refused, exactly as in
-//! [`super::v0`].
+//! Direction-aware and limit-lifting. Every input must agree on
+//! `left_to_right` — a conflict is a typed error instead of v0's
+//! silent drop — and the shared direction propagates to the merged
+//! root, so sub-level inputs no longer lose it under a synthesized
+//! root query. In addition, v1 merges *limited* path queries by
+//! **lifting**: an input's global `SizedQuery::limit` becomes its
+//! merged branch's per-instance cap (`Query::limit`) — exact, because
+//! the branch instance executes exactly once (its path is a concrete
+//! key chain), so "at most N rows from this whole input" and "at most
+//! N rows per instance" coincide — and authored per-instance limits
+//! ride along on their branches.
 //!
-//! Kept as its own dispatch arm (unlike the folded `path_query_push`
-//! intermediate) because it is selectable data: the `merge` slot is
-//! consulted by the verifier when re-deriving merged queries, and the
-//! 0/1/2 numbering is already spelled out in the `GROVE_V4` version
-//! table docs.
+//! Budgets never blend, so limits merge only as **exclusive grafts**:
+//!
+//! - a limited input whose whole path is the common path lands *at the
+//!   merged root*, where its query body would merge with the other
+//!   inputs' — refused;
+//! - two limited branches colliding on a key are refused;
+//! - limit-free inputs merge exactly as v0 does, modulo the direction
+//!   rules above, on the same code path.
+//!
+//! Offsets are still refused at every version.
+//!
+//! (An intermediate carrying only the direction rules was once gated
+//! here for `GROVE_V4`; since no grove version ever shipped it, it was
+//! folded into this v1 rather than kept as a dead dispatch arm —
+//! matching the `element.path_query_push` fold.)
 
 use grovedb_merk::proofs::query::{Query, QueryItem, SubqueryBranch};
 
-use crate::{Error, PathQuery};
+use crate::{operations::proof::util::hex_to_ascii, Error, PathQuery};
 
 /// `merge` v1 — see the module documentation.
 pub(super) fn merge_v1(path_queries: Vec<&PathQuery>) -> Result<PathQuery, Error> {
@@ -48,31 +61,47 @@ pub(super) fn merge_v1(path_queries: Vec<&PathQuery>) -> Result<PathQuery, Error
                 "can not merge pathqueries with offsets".to_string(),
             ));
         }
-        // Limits never merge here: a merged limit would silently mean
-        // something different than either input asked for. (Merge v2
-        // lifts them instead.)
-        if path_query.query.limit.is_some() {
-            return Err(Error::NotSupported(
-                "can not merge pathqueries with limits, consider setting the limit after the \
-                 merge"
-                    .to_string(),
-            ));
-        }
-        if path_query.has_instance_limits() {
-            return Err(Error::NotSupported(
-                "can not merge pathqueries carrying per-instance limits (Query::limit)".to_string(),
-            ));
-        }
+        let carries_limits = path_query.query.limit.is_some() || path_query.has_instance_limits();
         path_query
             .to_subquery_branch_with_offset_start_index(next_index)
-            .and_then(|unsized_path_query| {
+            .and_then(|mut unsized_path_query| {
                 if unsized_path_query.subquery_path.is_none() {
+                    // The input lands at the merged root, where its
+                    // query body merges with the other root-level
+                    // inputs — budgets cannot blend, so limits are
+                    // refused here even under v2.
+                    if carries_limits {
+                        return Err(Error::NotSupported(
+                            "can not merge a limited path query that lands at the merged root: \
+                             its budget would have to blend with the other queries' result \
+                             sets; give it a longer path of its own or set the limit after the \
+                             merge"
+                                .to_string(),
+                        ));
+                    }
                     queries_for_common_path_this_level.push(*unsized_path_query.subquery.ok_or(
                         Error::CorruptedCodeExecution(
                             "subquery must exist when subquery_path is none in merge",
                         ),
                     )?);
                 } else {
+                    // The lift: an input's global budget becomes its
+                    // branch-root query's per-instance cap. Exact,
+                    // because the branch instance executes exactly once
+                    // (its path is a concrete key chain), so "at most N
+                    // rows from this whole input" and "at most N rows
+                    // per instance" coincide.
+                    if let Some(global) = path_query.query.limit {
+                        let subquery = unsized_path_query.subquery.as_deref_mut().ok_or(
+                            Error::CorruptedCodeExecution(
+                                "subquery must exist on a sub-level merge branch",
+                            ),
+                        )?;
+                        subquery.limit = Some(match subquery.limit {
+                            Some(own) => own.min(global),
+                            None => global,
+                        });
+                    }
                     queries_for_common_path_sub_level.push(unsized_path_query);
                 }
                 Ok(())
@@ -101,11 +130,40 @@ pub(super) fn merge_v1(path_queries: Vec<&PathQuery>) -> Result<PathQuery, Error
             subquery_path: rest_of_path,
             subquery,
         };
-        // See v0: read modes are rejected in the prelude; propagate
-        // rather than discard if one ever reaches here.
-        merged_query
-            .merge_conditional_boxed_subquery(QueryItem::Key(key), subquery_branch)
-            .map_err(|e| Error::NotSupported(e.to_string()))?;
+        let limits_in_play = merged_query.has_instance_limit_anywhere()
+            || subquery_branch
+                .subquery
+                .as_deref()
+                .is_some_and(|subquery| subquery.has_instance_limit_anywhere());
+        if limits_in_play {
+            // Budgets never blend: a limit-carrying branch (lifted or
+            // authored) merges only as an exclusive graft. Any overlap
+            // with an existing conditional would need the two branches'
+            // bodies — and budgets — merged, which is refused by
+            // design.
+            let collides = merged_query
+                .conditional_subquery_branches
+                .as_ref()
+                .is_some_and(|branches| branches.keys().any(|item| item.contains(key.as_slice())));
+            if collides {
+                return Err(Error::NotSupported(format!(
+                    "can not merge limited path queries whose branches collide at key {}; \
+                     remove the limits or merge the colliding queries separately",
+                    hex_to_ascii(&key),
+                )));
+            }
+            merged_query.add_conditional_subquery(
+                QueryItem::Key(key),
+                subquery_branch.subquery_path,
+                subquery_branch.subquery.map(|subquery| *subquery),
+            );
+        } else {
+            // See v0: read modes are rejected in the prelude; propagate
+            // rather than discard if one ever reaches here.
+            merged_query
+                .merge_conditional_boxed_subquery(QueryItem::Key(key), subquery_branch)
+                .map_err(|e| Error::NotSupported(e.to_string()))?;
+        }
     }
 
     // The agreed direction travels to the merged root (it would
